Move connexion handler to handlers module for organiztion
